+++ src/Api/Strategies/Strategies.ts
@@ -1,36 +1,59 @@
-
-import { Strategy } from "../Types";
-import checkForSolved from "./checkForSolved";
-import hiddenPairsTriplesAndQuads from "./hiddenPairsTriplesAndQuads";
-import hiddenSingles from "./hiddenSingles";
-import intersectionRemoval from "./intersectionRemoval";
-import jellyfish from "./jellyfish";
-import pairsTriplesAndQuads from "./pairsTriplesAndQuads";
-import skyscraper from "./skyscraper";
-import swordfish from "./swordfish";
-import twoMinusOneLines from "./twoMinusOneLines";
-import updateCandidates from "./updateCandidates";
-import xWing from "./xWing";
-import xyChain from "./xyChain";
-import xyLoop from "./xyLoop";
-import yWing from "./yWing";
+import { Strategy } from '../Types'
 
 // See comments on `Strategy`
 const STRATEGIES = [
-   checkForSolved,
-   updateCandidates,
-   hiddenSingles,
-   intersectionRemoval,
-   pairsTriplesAndQuads,
-   hiddenPairsTriplesAndQuads,
-   xWing,
-   swordfish,
-   jellyfish,
-   skyscraper,
-   yWing,
-   twoMinusOneLines,
-   xyLoop,
-   xyChain,
-] as const
+   'checkForSolved',
+   'updateCandidates',
+   'hiddenSingles',
+   'intersectionRemoval',
+   'pairsTriplesAndQuads',
+   'hiddenPairsTriplesAndQuads',
+   'xWing',
+   'swordfish',
+   'jellyfish',
+   'skyscraper',
+   'yWing',
+   'twoMinusOneLines',
+   'xyLoop',
+   'xyChain',
+]
 
-export default STRATEGIES as typeof STRATEGIES & Strategy[]
+export const NUM_STRATEGIES = STRATEGIES.length
+
+/**
+ * Gets the strategy at an index\
+ * Used to lazyload
+ */
+export async function getStrategy<I extends number>(index: I): Promise<Strategy> {
+   if (index === 0) {
+      return (await import('./checkForSolved')).default
+   } else if (index === 1) {
+      return (await import('./updateCandidates')).default
+   } else if (index === 2) {
+      return (await import('./hiddenSingles')).default
+   } else if (index === 3) {
+      return (await import('./intersectionRemoval')).default
+   } else if (index === 4) {
+      return (await import('./pairsTriplesAndQuads')).default
+   } else if (index === 5) {
+      return (await import('./hiddenPairsTriplesAndQuads')).default
+   } else if (index === 6) {
+      return (await import('./xWing')).default
+   } else if (index === 7) {
+      return (await import('./swordfish')).default
+   } else if (index === 8) {
+      return (await import('./jellyfish')).default
+   } else if (index === 9) {
+      return (await import('./skyscraper')).default
+   } else if (index === 10) {
+      return (await import('./yWing')).default
+   } else if (index === 11) {
+      return (await import('./twoMinusOneLines')).default
+   } else if (index === 12) {
+      return (await import('./xyLoop')).default
+   } else if (index === 13) {
+      return (await import('./xyChain')).default
+   }
+
+   throw new RangeError(`Strategy index { ${index} } is invalid`)
+}