+++ src/Elems/StaticComponent.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default class StaticComponent extends React.PureComponent {
+   shouldComponentUpdate () { return false }
+}